change route defining structure
