+++ console/console-init/ui/src/Routes.tsx
@@ -15,6 +15,15 @@ const getAddressDetail = () => import("pages/AddressDetail/AddressDetailPage");
 const getConnectionDetail = () =>
   import("pages/ConnectionDetail/ConnectionDetailPage");
 
+const getConnectionsList = () =>
+  import(
+    "pages/AddressSpaceDetail/ConnectionList/ConnectionListWithFilterAndPaginationPage"
+  );
+const getAddressesList = () =>
+  import(
+    "pages/AddressSpaceDetail/AddressList/AddressesListWithFilterAndPaginationPage"
+  );
+
 export const AppRoutes = () => (
   <SwitchWith404>
     <Redirect path="/" to="/address-spaces" exact={true} />
@@ -36,6 +45,22 @@ export const AppRoutes = () => (
     <LazyRoute
       path="/address-spaces/:namespace/:name/:type/connections/:connectionname"
       getComponent={getConnectionDetail}
+      exact={true}
+    />
+  </SwitchWith404>
+);
+
+export const AddressSpaceRoutes = () => (
+  <SwitchWith404>
+    <Redirect path="/" to="/address-spaces" exact={true} />
+    <LazyRoute
+      path="/address-spaces/:namespace/:name/:type/addresses/"
+      getComponent={getAddressesList}
+      exact={true}
+    />
+    <LazyRoute
+      path="/address-spaces/:namespace/:name/:type/connections/"
+      getComponent={getConnectionsList}
       exact={true}
     />
   </SwitchWith404>
